consumers/requirements.txt
kafka-python
py7zr
nltk
scikit-learn
pandas
cassandra-driver
